requirements.txt
Django
python-dotenv
djangorestframework
django-cors-headers  
channels==3.0.5
daphne